refactor: rename AllComponentsContainer and restructure components for better organization
